libnvme: move fabrics-only helpers to util-fabrics.c
libnvmf_exat_ptr_next(), libnvmf_getifaddrs() (renamed from
libnvme_getifaddrs()), and the libnvmf_exat_len()/libnvmf_exat_size()
static inlines (renamed from nvmf_exat_len()/nvmf_exat_size()) are
only meaningful in fabrics-capable builds. Moving them out of util.c
and private.h into util-fabrics.c and private-fabrics.h keeps the
fabrics footprint contained and ensures PCIe-only/embedded builds do
not pull in this code.

util-fabrics.c is compiled only when want_fabrics is set, which
implies CONFIG_FABRICS is always defined, so no guard is needed around
libnvmf_getifaddrs() in that file.

While at it, simplify the util.c include guard from:
  #if defined(HAVE_NETDB) || defined(CONFIG_FABRICS)
to:
  #ifdef HAVE_NETDB
since CONFIG_FABRICS was only needed for libnvme_getifaddrs(), which
is now in util-fabrics.c.
